Move fastify module to struct
